Check and Clean zscan_fit
