--- z_scan/errorbar/__init__.py
@@ -1,3 +1,0 @@
-"""Package to compute the errorbars on the fitting results"""
-
-from .errorbar import compute
