--- z_scan/export_functions/__init__.py
@@ -1,7 +1,0 @@
-"""Package containing modules to export:
-- figures
-- text
-"""
-
-from export_functions import plot
-from export_functions import text
