--- z_scan/fitting_model/__init__.py
@@ -1,10 +1,0 @@
-"""Package containing packages and modules to run the fitting model.
-
-## Contains
----
-- run(): to run the fitting model
-- absorption: package to describe the absorption
-- chi2_minimising: package for the optimisation process
-"""
-
-from .run_model import run
