--- z_scan/fitting_model/absorption/__init__.py
@@ -1,10 +1,0 @@
-"""Package containing modules to describe the absorption
-
-## Contains
----
-- dI_dz: define the differential equation describing the absorption within a sample
-- intensity: return the beam intensity
-- transmittance: return the transmittance through a sample
-"""
-
-from .intensity import intensity
