--- z_scan/fitting_model/chi2_minimising/__init__.py
@@ -1,1 +1,0 @@
-"""Package for the optimisation process"""
