+++ zscan/__init__.py
@@ -1,0 +1,5 @@
+"""Package containing several functionalities necessary for fitting z-scan data"""
+
+import zscan.general
+from zscan.errorbar import compute as errorbar
+from zscan.run_model import run
