+++ zscan/optimize/__init__.py
@@ -1,0 +1,4 @@
+"""Package for the optimisation process"""
+
+import zscan.optimize.basinhopping
+import zscan.optimize.chi2
